+++ FBSDKGamingServicesKit/FBSDKGamingServicesKitTests/SwitchContextDialogTests.swift
@@ -178,32 +178,6 @@ final class SwitchContextDialogTests: XCTestCase, ContextDialogDelegate {
     XCTAssertNotNil(dialogError)
   }
 
-  // MARK: - Deprecated methods from former wrapper
-
-  func testCreatingWithFactoryMethod() {
-    dialog = SwitchContextDialog.dialog(
-      content: content,
-      windowFinder: windowFinder,
-      delegate: self
-    )
-
-    XCTAssertIdentical(
-      dialog.dialogContent,
-      content,
-      "The dialog should be created with the provided content"
-    )
-    XCTAssertIdentical(
-      dialog.windowFinder,
-      windowFinder,
-      "The dialog should be created with the provided window finder"
-    )
-    XCTAssertIdentical(
-      dialog.delegate,
-      self,
-      "The dialog should be created with the provided delegate"
-    )
-  }
-
   // MARK: - Delegate Methods
 
   func contextDialogDidComplete(_ contextDialog: ContextWebDialog) {